+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 Changelog
 =========
 
+3.2.0 (2021-03-11)
+-------------------
+* [ADD] Flattener Transformer (explodes and de-nests complex dataframes)
+* [MOD] add ignore_ambiguous_columns to Mapper (and Flattener)
+* [MOD] log spooq version when importing
+* [REM] Drop separate spark package (bin-folder) as pip package can now handle all tests as well
+
 3.1.0 (2021-01-27)
 -------------------
 * [ADD] EnumCleaner Transformer